Cria componentes dedicados
