+++ packages/byte-bank/src/app/components/Dashboard/Balance/index.ts
@@ -1,0 +1,2 @@
+export * from "./Balance";
+export { default } from "./Balance";
