+++ packages/byte-bank/src/app/components/Dashboard/Extract/index.ts
@@ -1,0 +1,2 @@
+export * from "./Extract";
+export { default } from "./Extract";
