+++ packages/byte-bank/src/app/components/Dashboard/Transactions/index.ts
@@ -1,0 +1,2 @@
+export * from "./Transactions";
+export { default } from "./Transactions";
