+++ packages/byte-bank/src/app/components/index.ts
@@ -1,1 +1,4 @@
 export {default as Header} from './Header/index';
+export {default as Balance} from './Dashboard/Balance';
+export {default as Extract} from './Dashboard/Extract';
+export {default as Transactions} from './Dashboard/Transactions';
